refactor flatten summary from datasetprofile
